Factor out voting, inversion, and util functions
