+++ torchfields/fields.py
@@ -1,101 +1,18 @@
 """PyTorch tensor type for working with displacement vector fields
 """
 import torch
-import torch.nn as nn
 import torch.nn.functional as F
 from functools import wraps
 
-
-#############################################
-# Decorators for enforcing return value types
-#############################################
-
-
-def return_subclass_type(cls):
-    """Class decorator for a subclass to encourage it to return its own
-    subclass type whenever its inherited functions would otherwise return
-    the superclass type.
-
-    This works by attempting to convert any return values of the superclass
-    type to the subclass type, and then defaulting back to the original
-    return value on any errors during conversion.
-
-    If running the subclass constructor has undesired side effects,
-    the class can define a `_from_superclass()` function that casts
-    to the subclass type more directly.
-    This function should raise an exception if the type is not compatible.
-    If `_from_superclass` is not defined, the class constructor is called
-    by default.
-    """
-    def decorator(f):
-        @wraps(f)
-        def f_decorated(*args, **kwargs):
-            out = f(*args, **kwargs)
-            try:
-                if not isinstance(out, cls) and isinstance(out, cls.__bases__):
-                    return cls._from_superclass(out)
-            except Exception:
-                pass
-            # Result cannot be returned as subclass type
-            return out
-        return f_decorated
-
-    # fall back to constructor if _from_superclass not defined
-    try:
-        cls._from_superclass
-    except AttributeError:
-        cls._from_superclass = cls
-
-    for name in dir(cls):
-        attr = getattr(cls, name)
-        if name not in dir(object) and callable(attr):
-            try:
-                # check if this attribute is flagged to keep its return type
-                if attr._keep_type:
-                    continue
-            except AttributeError:
-                pass
-            setattr(cls, name, decorator(attr))
-    return cls
-
-
-def dec_keep_type(keep=True):
-    """Function decorator that adds a flag to tell `return_subclass_type()`
-    to leave the function's return type as is.
-
-    This is useful for functions that intentionally return a value of
-    superclass type.
-
-    If a boolean argument is passed to the decorator as
-
-        @dec_keep_type(True)
-        def func():
-            pass
-
-    then that agument determines whether to enable the flag. If no argument
-    is passed, the flag is enabled as if `True` were passed.
-
-        @dec_keep_type
-        def func():
-            pass
-
-    """
-    def _dec_keep_type(keep_type):
-        def _set_flag(f):
-            f._keep_type = keep_type
-            return f
-        return _set_flag
-    if isinstance(keep, bool):  # boolean argument passed
-        return _dec_keep_type(keep)
-    else:  # the argument is actually the function itself
-        func = keep
-        return _dec_keep_type(True)(func)
+from .utils import (return_subclass_type, dec_keep_type, permute_input,
+                    permute_output, ensure_dimensions)
+from . import inversion
+from . import voting
 
 
 ####################################
 # DisplacementField Class Definition
 ####################################
-
 
 @return_subclass_type
 class DisplacementField(torch.Tensor):
@@ -186,81 +103,6 @@ class DisplacementField(torch.Tensor):
         """Converts the `DisplacementField` to a standard `torch.Tensor`
         """
         return self.clone().tensor_()
-
-    # Decorators to convert the inputs and outputs of DisplacementField methods
-
-    def permute_input(f):
-        """Function decorator to permute the input dimensions from the
-        DisplacementField convention `(N, 2, H, W)` to the standard PyTorch
-        field convention `(N, H, W, 2)` before passing it into the function.
-        """
-        @wraps(f)
-        def f_new(self, *args, **kwargs):
-            ndims = self.ndimension()
-            perm = self.permute(*range(ndims-3), -2, -1, -3)
-            return f(perm, *args, **kwargs)
-        return f_new
-
-    def permute_output(f):
-        """Function decorator to permute the dimensions of the function output
-        from the standard PyTorch field convention `(N, H, W, 2)` to the
-        DisplacementField convention `(N, 2, H, W)` before returning it.
-        """
-        @wraps(f)
-        def f_new(self, *args, **kwargs):
-            out = f(self, *args, **kwargs)
-            ndims = out.ndimension()
-            return out.permute(*range(ndims-3), -1, -3, -2)
-        return f_new
-
-    def ensure_dimensions(ndimensions=4, arg_indices=(0,), reverse=False):
-        """Function decorator to ensure that the the input has the
-        approprate number of dimensions
-
-        If it has too few dimensions, it pads the input with dummy dimensions.
-
-        Args:
-            ndimensions (int): number of dimensions to pad to
-            arg_indices (int or List[int]): the indices of inputs to pad
-                Note: Currently, this only works on arguments passed by
-                position. Those inputs must be a torch.Tensor or
-                DisplacementField.
-            reverse (bool): if `True`, it then also removes the added dummy
-                dimensions from the output, down to the number of dimensions
-                of arg[arg_indices[0]]
-        """
-        if callable(ndimensions):  # it was called directly on a function
-            func = ndimensions
-            ndimensions = 4
-        else:
-            func = None
-        if isinstance(arg_indices, int):
-            arg_indices = (arg_indices,)
-        assert(len(arg_indices) > 0)
-
-        def decorator(f):
-            @wraps(f)
-            def f_decorated(*args, **kwargs):
-                args = list(args)
-                original_ndims = len(args[arg_indices[0]].shape)
-                for i in arg_indices:
-                    if i >= len(args):
-                        continue
-                    while args[i].ndimension() < ndimensions:
-                        args[i] = args[i].unsqueeze(0)
-                out = f(*args, **kwargs)
-                while reverse and out.ndimension() > original_ndims:
-                    new_out = out.squeeze(0)
-                    if new_out.ndimension() == out.ndimension():
-                        break  # no progress made; nothing left to squeeze
-                    out = new_out
-                return out
-            return f_decorated
-
-        if func is None:  # parameters were passed to the decorator
-            return decorator
-        else:  # the function itself was passed to the decorator
-            return decorator(func)
 
     # Constuctors for typical displacent fields
 
@@ -1031,258 +873,13 @@ class DisplacementField(torch.Tensor):
         """
         return self.inverse(*args, **kwargs)
 
-    @ensure_dimensions(ndimensions=4, arg_indices=(0), reverse=True)
+    @wraps(inversion.linverse)
     def linverse(self, autopad=True):
-        """Return a left inverse approximation for the displacement field
+        return inversion.linverse(self, autopad=True)
 
-        Given a displacement field `f`, its left inverse is a displacement
-        field `g` such that
-        `g(f) ~= identity`
-
-        In other words
-        :math:`f_{inv} = \argmin_{g} |g(f)|^2`
-        """
-        if len(self.shape) != 4 or self.shape[0] > 1:
-            raise NotImplementedError('Left inverse is currently implemented '
-                                      'only for single-batch fields. '
-                                      'Received batch size {}'
-                                      .format(','.join(
-                                          str(n) for n in self.shape[:-3])))
-        # comparison to 0
-        eps1 = 2**(-51) if self.dtype is torch.double else 2**(-23)
-        # denominator fudge factor to avoid dividing by 0
-        eps2 = eps1 * 2**(-10)
-        # tolarance for point containment in a quadrilateral
-        eps3 = 2**-16
-
-        def tensor_min(*args):
-            """Elementwise minimum of a sequence of tensors"""
-            minimum, *rest = args
-            for arg in rest:
-                minimum = minimum.min(arg)
-            return minimum
-
-        def tensor_max(*args):
-            """Elementwise maximum of a sequence of tensors"""
-            maximum, *rest = args
-            for arg in rest:
-                maximum = maximum.max(arg)
-            return maximum
-
-        def pad(inp):
-            """Pads the field just enough to eliminate border effects"""
-            with torch.no_grad():
-                *_, H, W = inp.shape
-                mapping = inp.pixel_mapping()
-                pad_yl = mapping.y[..., 0, :].max().ceil().int().item()
-                pad_yh = (H-1-mapping.y[..., -1, :].min()).ceil().int().item()
-                pad_xl = mapping.x[..., :, 0].max().ceil().int().item()
-                pad_xh = (W-1-mapping.x[..., :, -1].min()).ceil().int().item()
-                pad_yl = max(pad_yl, 0) + 1
-                pad_yh = max(pad_yh, 0) + 1
-                pad_xl = max(pad_xl, 0) + 1
-                pad_xh = max(pad_xh, 0) + 1
-                # ensure that the new field is square (that is, newH = newW)
-                newH, newW = pad_yl + H + pad_yh, pad_xl + W + pad_xh
-                if newH > newW:
-                    pad_xh += newH - newW
-                elif newW > newH:
-                    pad_yh += newW - newH
-                padding = (pad_xl, pad_xh, pad_yl, pad_yh)
-            return (F.pad(inp.pixels(), padding, mode='replicate').field()
-                    .from_pixels(), padding)
-
-        def unpad(inp, padding):
-            """Crops the field back to its original size"""
-            p_xl, p_xh, p_yl, p_yh = padding
-            p_xh = inp.shape[-1] - p_xh
-            p_yh = inp.shape[-2] - p_yh
-            return inp.pixels()[..., p_yl:p_yh, p_xl:p_xh].from_pixels()
-
-        def fold(inp):
-            """Collapse the matrix at each pixel onto the local neighborhood
-
-            The input to this function is a spatial tensor in which the
-            entry contained in every spatial pixel is itself a small matrix.
-            The values of this matrix correspond to the neighborhood of that
-            pixel. For instance, the value at the center of the matrix
-            corresponds to the pixel itself, whereas the value above and to
-            the left of the center corresponds to the pixel's upper left
-            neighbor, and so on.
-
-            This function collapses this into a spatial tensor with scalar
-            values by summing the respective values corresponding to each
-            pixel.
-            """
-            pad = (0, (inp.shape[2] + 1) % 2,  # last dimension
-                   0, (inp.shape[1] + 1) % 2)  # second to last dimension
-            res = F.pad(inp, pad)
-            res = F.fold(
-                res.view(1, res.shape[0]*res.shape[1]*res.shape[2], -1),
-                output_size=inp.shape[3:], kernel_size=inp.shape[1:3],
-                padding=((inp.shape[1])//2, (inp.shape[2])//2))
-            return res
-
-        @torch.no_grad()
-        def winding_number(Px, Py, v00, v01, v10, v11):
-            """Gives the winding number of a quadrilateral around a grid point
-            When the winding number is non-zero, the quadrilateral contains
-            the grid point. More specifically, only positive winding numbers
-            are relevant in this context, since inverted quadrilaterals are
-            not considered.
-            For edge cases, this is a more accurate measure than checking
-            whether i and j fall within the range [0,1).
-            Based on http://geomalgorithms.com/a03-_inclusion.html
-            """
-            try:
-                # vertices in counterclockwise order (viewing y as up)
-                V = torch.stack([v00, v01, v11, v10, v00], dim=0).field_()
-                v00 = v01 = v10 = v11 = None  # del v00, v01, v10, v11
-                # initial and final vertex for each edge
-                Vi, Vf = V[:-1], V[1:]
-                V = None  # del V
-                # sign of cross product indicates direction around grid point
-                cross = ((Vf.x - Vi.x)*(Py - Vi.y) - (Px - Vi.x)*(Vf.y - Vi.y))
-                # upward crossing of rightward ray from grid point
-                upward = (Vi.y <= Py) & (Vf.y > Py) & (cross > -eps3)
-                # downward crossing of rightward ray from grid point
-                downward = (Vi.y > Py) & (Vf.y <= Py) & (cross < -eps3)
-                Vi = Vf = Px = Py = cross = None  # del Vi, Vf, Px, Py, cross
-                # winding number = diff between number of up and down crossings
-                return (upward.int() - downward.int()).sum(dim=0)
-            except RuntimeError:
-                # In case this is an out-of-memory error, clear temp tensors
-                v00 = v01 = v10 = v11 = Px = Py = None
-                V = Vi = Vf = cross = upward = downward = None
-                raise
-
-        try:
-            # pad the field
-            if autopad:
-                field, padding = pad(self)
-            else:
-                field = self
-                padding = None
-
-            # vectors at the four corners of each pixel's quadrilateral
-            mapping = field.pixel_mapping()
-            v00 = mapping[..., :-1, :-1]
-            v01 = mapping[..., :-1, 1:]
-            v10 = mapping[..., 1:, :-1]
-            v11 = mapping[..., 1:, 1:]
-            mapping = None  # del mapping
-
-            with torch.no_grad():
-                # find each quadrilateral's (set of 4 vectors) span, in pixels
-                v_min = tensor_min(v00, v01, v10, v11).floor()
-                v_min.y.clamp_(0, field.shape[-2] - 1)
-                v_min.x.clamp_(0, field.shape[-1] - 1)
-                v_max = tensor_max(v00, v01, v10, v11).ceil()
-                v_max.y.clamp_(0, field.shape[-2] - 1)
-                v_max.x.clamp_(0, field.shape[-1] - 1)
-                # d_x and d_y are the largest spans in x and y
-                d = (v_max - v_min).max_vector().max(0)[0].int()
-                v_max = None  # del v_max
-                d_x, d_y = list(d.cpu().numpy())
-                d = ((d//2).unsqueeze(-1).unsqueeze(-1)  # center of the span
-                     .unsqueeze(-1).unsqueeze(-1)).to(v_min)
-                v_min.y.clamp_(0, field.shape[-2] - 1 - d_y)
-                v_min.x.clamp_(0, field.shape[-1] - 1 - d_x)
-                # u is an identity pixel mapping of a d_y by d_x neighborhood
-                u = field.identity().pixel_mapping()[..., :d_y, :d_x].round()
-                ux = u.x.unsqueeze(-1).unsqueeze(-1)
-                uy = u.y.unsqueeze(-1).unsqueeze(-1)
-                u = None  # del u
-
-            # subtract out v_min to bring all quadrilaterals near zero
-            v00 = (v00 - v_min).unsqueeze(-4).unsqueeze(-4)
-            v01 = (v01 - v_min).unsqueeze(-4).unsqueeze(-4)
-            v10 = (v10 - v_min).unsqueeze(-4).unsqueeze(-4)
-            v11 = (v11 - v_min).unsqueeze(-4).unsqueeze(-4)
-
-            # quadratic coefficients in gridsample solution `a*j^2+b*j+c=0`
-            a = ((v00.x - v01.x) * (v00.y - v01.y - v10.y + v11.y)
-                 - (v00.x - v01.x - v10.x + v11.x) * (v00.y - v01.y))
-            b = ((ux - v00.x) * (v00.y - v01.y - v10.y + v11.y)
-                 + (v00.x - v01.x) * (-v00.y + v10.y)
-                 - (-v00.x + v10.x) * (v00.y - v01.y)
-                 - (v00.x - v01.x - v10.x + v11.x) * (uy - v00.y))
-            c = (ux - v00.x)*(-v00.y + v10.y) - (-v00.x + v10.x)*(uy - v00.y)
-            # quadratic formula solution (note positive root is always invalid)
-            j_temp = ((b + (b.pow(2) - 4*a*c).clamp(min=eps2).sqrt()).abs()
-                      / ((2*a).abs() + eps2))
-            # corner case when a == 0 (reduces to `b*j + c = 0`)
-            j_temp = j_temp.where(a.abs() > eps1, -c / b)
-            a = b = c = None  # del a, b, c
-            # get i from j_temp
-            i = ((uy - v00.y + (v00.y - v01.y) * j_temp).abs()
-                 / ((-v00.y + v10.y + (v00.y - v01.y - v10.y + v11.y) * j_temp)
-                    .abs() + eps2))
-            j_temp = None  # del j_temp
-            # j has significantly smaller rounding error for near-trapezoids
-            j = ((ux - v00.x + (v00.x - v10.x) * i).abs()
-                 / ((-v00.x + v01.x + (v00.x - v10.x - v01.x + v11.x) * i)
-                    .abs() + eps2))
-            # winding_number > 0 means point is contained in the quadrilateral
-            wn = winding_number(ux, uy, v00, v01, v10, v11)
-            ux = uy = None  # del ux, uy
-            v00 = v01 = v10 = v11 = None  # del v00, v01, v10, v11
-
-            # negative of the bilinear interpolation to produce inverse vector
-            v00 = field[..., :-1, :-1].unsqueeze(-3).unsqueeze(-3)
-            v01 = field[..., :-1, 1:].unsqueeze(-3).unsqueeze(-3)
-            v10 = field[..., 1:, :-1].unsqueeze(-3).unsqueeze(-3)
-            v11 = field[..., 1:, 1:].unsqueeze(-3).unsqueeze(-3)
-            inv = -((1-i)*(1-j)*v00 + (1-i)*j*v01 + i*(1-j)*v10 + i*j*v11)
-            v00 = v01 = v10 = v11 = None  # del v00, v01, v10, v11
-
-            # mask out inverse vectors at points outside the quadrilaterals
-            mask = (wn > 0) & torch.isfinite(i) & torch.isfinite(j)
-            i = j = wn = None  # del i, j, wn
-            inv = inv.where(mask, torch.tensor(0.).to(inv))
-            # append mask to keep track of how many contributions to each pixel
-            inv = torch.cat((inv, mask.to(inv)), 1)
-            mask = None
-
-            # indices at which to place each inverse vector in a sparse tensor
-            indices = ((v_min.unsqueeze(-3).unsqueeze(-3) + d)
-                       .view(2, -1).flip(0).round().long())
-            v_min = d = None
-            inv = inv.view(3, d_y, d_x, -1).permute(3, 0, 1, 2)
-            # construct sparse tensor and use `to_dense` to arrange vectors
-            SparseTensor = (torch.cuda.sparse.FloatTensor if self.is_cuda
-                            else torch.sparse.FloatTensor)
-            inv = SparseTensor(indices, inv, (*field.shape[-2:], 3, d_y, d_x),
-                               device=inv.device)
-            inv = inv.to_dense().permute(2, 3, 4, 0, 1)
-            # fold the d_y by d_x neighborhoods by summing the overlaps
-            inv = fold(inv)
-            # divide each pixel by number of contributions to get an average
-            inv = inv[:, :2] / inv[:, 2:].clamp(min=1.)
-
-            # crop back to original shape
-            if autopad:
-                inv = unpad(inv.field(), padding)
-            return inv
-        except RuntimeError:
-            # In case this is an out-of-memory error, clear temporary tensors
-            self = field = pad = mapping = v_min = v_max = d = d_x = d_y = None
-            u = ux = uy = v00 = v01 = v10 = v11 = wn = None
-            a = b = c = j_temp = i = j = None
-            mask = inv = indices = None
-            raise
-
+    @wraps(inversion.rinverse)
     def rinverse(self, *args, **kwargs):
-        """Return a right inverse approximation for the displacement field
-
-        Given a displacement field `f`, its right inverse is a displacement
-        field `g` such that
-        `f(g) ~= identity`
-
-        In other words
-        :math:`f_{inv} = \argmin_{g} |f(g)|^2`
-        """
-        raise NotImplementedError
+        return inversion.rinverse(self, autopad=True)
 
     # Adapting functions inherited from torch.Tensor
 
@@ -1311,86 +908,10 @@ class DisplacementField(torch.Tensor):
 
     # Vector Voting
 
+    @wraps(voting.gaussian_blur)
     def gaussian_blur(self, sigma=1, kernel_size=5):
-        """Gausssian blur the displacement field to reduce any unsmoothness
-        Adapted from https://bit.ly/2JO7CCP
-        """
-        import math
-        if sigma == 0:
-            return self.clone()
-        pad = (kernel_size - 1) // 2
-        if kernel_size % 2 == 0:
-            pad = (pad, pad+1, pad, pad+1)
-        else:
-            pad = (pad,)*4
-        padded = F.pad(self, pad, mode='reflect')
-        mu = (kernel_size - 1) / 2
-        x = torch.stack(torch.meshgrid([torch.arange(kernel_size).to(self)]*2))
-        kernel = torch.exp((-((x - mu) / sigma) ** 2) / 2)
-        kernel = kernel.prod(dim=0) / (2 * math.pi * sigma**2)
-        kernel = kernel / kernel.sum()  # renormalize to get unit product
-        kernel = kernel.expand(2, 1, *kernel.shape)
-        return F.conv2d(padded, weight=kernel, groups=2)
+        return voting.gaussian_blur(self, sigma, kernel_size)
 
+    @wraps(voting.vote)
     def vote(self, softmin_temp=1, blur_sigma=1):
-        """Produce a single, consensus displacement field from a batch of
-        displacement fields
-
-        The resulting displacement field represents displacements that are
-        closest to the most consistent majority of the fields.
-        This effectively allows the fields to differentiably vote on the
-        displacement that is most likely to be correct.
-
-        Args:
-            self: DisplacementField of shape (N, 2, H, W)
-            softmin_temp (float): temperature of softmin to use
-            blur_sigma (float): std dev of the Gaussian kernel by which to blur
-                the softmin inputs. Note that the outputs are not blurred.
-                None or 0 means no blurring.
-
-        Returns:
-            DisplacementField of shape (1, 2, H, W) containing the vector
-            voting result
-        """
-        from itertools import combinations
-        if self.ndimension() != 4:
-            raise ValueError('Vector voting is only implemented on '
-                             'displacement fields with 4 dimensions. '
-                             'The input has {}.'.format(self.ndimension()))
-        n, _two_, *shape = self.shape
-        if n == 1:
-            return self
-        elif n % 2 == 0:
-            raise ValueError('Cannot vetor vote on an even number of '
-                             'displacement fields: {}'.format(n))
-        m = (n + 1) // 2  # smallest number that constututes a majority
-        blurred = self.gaussian_blur(sigma=blur_sigma) if blur_sigma else self
-
-        # compute distances for all pairs of fields
-        dists = torch.zeros((n, n, *shape)).to(device=blurred.device)
-        for i in range(n):
-            for j in range(i):
-                dists[i, j] = dists[j, i] \
-                    = blurred[i].distance(blurred[j])
-
-        # compute mean distance for all m-tuples
-        mtuples = list(combinations(range(n), m))
-        mtuple_avg = []
-        for mtuple in mtuples:
-            delta = torch.stack([
-                dists[i, j] for i, j in combinations(mtuple, 2)
-            ]).mean(dim=0)
-            mtuple_avg.append(delta)
-        mavg = torch.stack(mtuple_avg)
-
-        # compute weights for mtuples: smaller mean distance -> higher weight
-        mt_weights = (-mavg/softmin_temp).softmax(dim=0)
-
-        # assign mtuple weights back to individual fields
-        field_weights = torch.zeros((n, *shape)).to(device=mt_weights.device)
-        for i, mtuple in enumerate(mtuples):
-            for j in mtuple:
-                field_weights[j] += mt_weights[i]
-        field_weights = field_weights / field_weights.sum(dim=0, keepdim=True)
-
-        return (self * field_weights.unsqueeze(-3)).sum(dim=0, keepdim=True)
+        return voting.vote(self, softmin_temp, blur_sigma)
